requirements.txt
streamlit==1.32.2
openpyxl==3.1.2
openai>=2.0.0
tenacity==8.2.3
python-dotenv==1.0.1
pandas==2.2.1
为了方便你测试，我在 [db.py](file:///Users/ricardo/文稿/创业/软件服务脚本/excel翻译/auth/db.py#L29-34) 中预设了两个测试卡密。你可以复制以下任意一个使用：

1. **次卡（推荐测试用）**：
   - 卡密：`TEST-COUNT-100`
   - 说明：包含 100 次翻译额度，永久有效。

2. **月卡（测试时间限制用）**：
   - 卡密：`TEST-MONTH-1`
   - 说明：有效期一个月，每天重置 1000 次额度。

你可以直接复制 `TEST-COUNT-100` 填入左侧的输入框点击激活即可！